requirements.txt
pandas
numpy
matplotlib
seaborn
nltk
gensim
emoji
praw
tweepy
progressbar2
lda